client/requirements.txt
websocket
InquirerPy
ezcord